Fixing schema to support allowing querying other fields alongside viterbi fields
Summary: Previously, viterbi queries were passed straight to viterbi, as they were. This could lead to some fields that were meant to be querying the cwfollowup schema being passed to viterbi, with undesirable results. This is now fixed by reconstructing a query specifically for viterbi

Test Plan: None yet, but this is coming once the main cwfollowup scaffolding is in place.

Reviewers: llakerink, mr-asher

Reviewed By: mr-asher

Differential Revision: https://phab.adacs.org.au/D695